Optimize Multiplatform cache builds
This PR starts an ARM EC2 instance and forwards socket via SSH
so that it can be used by docker buildx build with airflow_cache
multi-platform builders. Then it appends it to the regular
builder making the builds run up to 10 minutes rather than
1.5 hours.

The instance is created only when "main" build reaches the cache
build job/step.

The instances run for maximum 50 minutes and then self-terminate just
in case, also the instance is killed immediately when the job finishes
(no matter if succesfully or not).